Fix navigation when mod is expanded
